+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 - Added `Sample::bits_per_sample` method.
 - ALSA(process_output): Pass `silent=true` to `PCM.try_recover`, so it doesn't write to stderr.
-- ALSA: Fix buffer and period size by selecting the closest supported values.
-- ALSA: Change ALSA periods from 4 to 2.
+- ALSA: Fix BufferSize::Fixed by selecting the nearest supported frame count.
+- ALSA: Change BufferSize::Default to use the device defaults instead of 4 periods of 25 ms.
 - ALSA: Change card enumeration to work like `aplay -L` does.
 - ASIO: Fix linker flags for MinGW cross-compilation.
 - CoreAudio: Change `Device::supported_configs` to return a single element containing the available sample rate range when all elements have the same `mMinimum` and `mMaximum` values.